+++ src/maxgradient/__init__.py
@@ -35,7 +35,7 @@ from maxgradient.rule import GradientRule
 from maxgradient.spectrum import Spectrum
 from maxgradient.theme import GRADIENT_TERMINAL_THEME, GradientTheme
 
-__version__ = "0.2.19"
+__version__ = "0.3.0"
 
 __all__ = [
     "Color",